ether-state/x/etherstate/client/cli/query_ethereum_address.go
package cli

import (
	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/spf13/cobra"

	"github.com/nelsonstr/o3n1/ether-state/x/etherstate/types"
)

func CmdListEthereumAddress() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "list-ethereum-address",
		Short: "list all ethereum-address",
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx, err := client.GetClientQueryContext(cmd)
			if err != nil {
				return err
			}

			pageReq, err := client.ReadPageRequest(cmd.Flags())
			if err != nil {
				return err
			}

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryAllEthereumAddressRequest{
				Pagination: pageReq,
			}

			res, err := queryClient.EthereumAddressAll(cmd.Context(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}

func CmdShowEthereumAddress() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-ethereum-address [index]",
		Short: "shows a ethereum-address",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) (err error) {
			clientCtx, err := client.GetClientQueryContext(cmd)
			if err != nil {
				return err
			}

			queryClient := types.NewQueryClient(clientCtx)

			argIndex := args[0]

			params := &types.QueryGetEthereumAddressRequest{
				Index: argIndex,
			}

			res, err := queryClient.EthereumAddress(cmd.Context(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
